rework
decided to readd json version, it will work if the driver profiles are stored directly on the roborio. this does make sense, but it will be hard to write to those files or transfer them between rios (use the same SD card?).
we can also just access the roborio and SCP those files over.

also renamed SwerveConstants to SwerveSettings because some things are constant anymore..